refactor: service and app router
